fix(test): move new tests to existing test file
- Remove duplicate test file at tests/unit/scripts/
- Add TestCheckRedactionFailModeDocs to tests/unit/test_check_doc_accuracy.py
- Fixes CI module import conflict
